docs: document controller layer
