+++ src/main/java/edu/kit/hci/soli/controller/ErrorController.java
@@ -13,12 +13,27 @@ import java.util.Map;
 
 import static org.springframework.boot.web.error.ErrorAttributeOptions.Include.*;
 
+/**
+ * Controller for handling errors in the application.
+ */
 @Controller
 public class ErrorController extends AbstractErrorController {
+    /**
+     * Constructs an ErrorController with the specified {@link DefaultErrorAttributes}.
+     *
+     * @param errorAttributes the default error attributes
+     */
     public ErrorController(DefaultErrorAttributes errorAttributes) {
         super(errorAttributes);
     }
 
+    /**
+     * Handles errors and returns the appropriate error view.
+     *
+     * @param model the model to be used in the view
+     * @param request the HTTP request
+     * @return the view name
+     */
     @RequestMapping("/error")
     public String handleError(Model model, HttpServletRequest request) {
         Map<String, Object> errorAttributes = getErrorAttributes(request, ErrorAttributeOptions.of(
