+++ src/main/java/edu/kit/hci/soli/controller/UsersController.java
@@ -11,15 +11,32 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+/**
+ * Controller class for managing user-related operations.
+ */
 @Controller
 @Slf4j
 public class UsersController {
     private final UserService userService;
 
+    /**
+     * Constructs a UsersController with the specified {@link UserService}.
+     *
+     * @param userService the service for managing users
+     */
     public UsersController(UserService userService) {
         this.userService = userService;
     }
 
+    /**
+     * Deactivates a user by their ID.
+     *
+     * @param model the model to be used in the view
+     * @param response the HTTP response
+     * @param principal the authenticated user details
+     * @param userId the ID of the user to be deactivated
+     * @return the view name
+     */
     @PutMapping("/admin/users/{userId}/deactivate")
     public String deactivateUser(Model model, HttpServletResponse response, @AuthenticationPrincipal SoliUserDetails principal, @PathVariable Long userId) {
         log.info("User {} requested to deactivate user {}", principal.getUser(), userId);
@@ -43,11 +60,25 @@ public class UsersController {
         return "users";
     }
 
+    /**
+     * Returns the view for disabled users.
+     *
+     * @param model the model to be used in the view
+     * @return the view name
+     */
     @RequestMapping("/disabled")
     public String getDisabled(Model model) {
         return "disabled_user";
     }
 
+    /**
+     * Retrieves all manageable users and returns the view for the users page.
+     *
+     * @param model the model to be used in the view
+     * @param response the HTTP response
+     * @param principal the authenticated user details
+     * @return the view name
+     */
     @GetMapping("/admin/users")
     public String getUsers(Model model, HttpServletResponse response, @AuthenticationPrincipal SoliUserDetails principal) {
         log.info("User {} requested the users page", principal.getUser());
