+++ src/main/java/edu/kit/hci/soli/controller/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains the controllers for the Soli application.
+ * Controllers handle HTTP requests and responses, and interact with the service layer to process data.
+ */
+package edu.kit.hci.soli.controller;
